refactor(server): improve stderr logging and refine configuration

+++ rust/multilspy-rust/src/server.rs
@@ -107,7 +107,10 @@ fn spawn_lsp_server_process(
 
     cmd.stdout(Stdio::piped()).stdin(Stdio::piped());
 
-    if let Some(path) = &config.ra_stderr_log_path {
+    if let Some(path) = &config.ra_stderr_log_path
+        && path.exists()
+        && path.is_file()
+    {
         cmd.stderr(std::fs::File::create(path)?);
     } else {
         cmd.stderr(Stdio::piped());
@@ -641,8 +644,7 @@ mod tests {
         )
         .with_stderr_log_path(PathBuf::from("./ra_stderr.log"))
         .with_env("RA_LOG".to_string(), "info".to_string())
-        .with_wait_work_done_progress_create_max_time(Duration::from_secs(5))
-        ;
+        .with_wait_work_done_progress_create_max_time(Duration::from_secs(5));
 
         let server = RustAnalyzerServer::start_server(config).await.unwrap();
